requirements.txt.txt
asgiref==3.6.0
crispy-bootstrap4==2022.1
Django==4.1.7
django-crispy-forms==2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.4.0
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
